Ditch async batch processing in favour of serially processing the songs in the background
I ALMOST had the async analysis working, but there were too many race conditions when trying to synchronize
everything when saving contexts to the persistent store. Trying to debug this was getting close to scope
creep territory so I'm abandoning the multithreading explorations. Maybe a rainy-day project later in the
summer