Add a GrandSelektor class to handle track selection
This class will be the meat and potatoes of the application, in charge
of selecting the next song given a currently-playing song. Build-time
algorithm specification can be done by specifying the `selektorAlgorithm`
key in the Settings.plist file, which will change the algorithm the
GrandSelektor uses to pick the next song. For now, I'm using a 'dummy'
algorithm which simply picks a track at random. TODO: Implement the
algorithms themselves, prevent the current song from being selected, and
prevent un-analyzed songs from being selected